sentiment/requirements.txt
# Sentiment-Aware Forex Monitor - Dependencies

# Core
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
feedparser>=6.0.10

# Web Framework
Flask>=3.0.0
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL support (falls back to SQLite if not available)
gspread>=5.0.0  # Google Sheets support (preferred option)
google-auth>=2.0.0  # Google authentication for Sheets API

# Data & Analysis
yfinance>=0.2.0

# LLM Providers (install the ones you need)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Scheduling
schedule>=1.2.0

# Email
# (uses built-in smtplib, no additional package needed)

